Visual updates, food blog introduced
Lots and lots of little visual tweaks.
First and foremost, there's a new wallpaper.
Secondly, the color scheme for most of the site has been updated to a
nice, cool green. This should be a welcome change to anyone who (like
me) didn't care much for the orange-gold design of the old version of
the site.

The FOOD BLOG!
I like food. Therefore, I've decided to talk about it.

That's about it.