New blog sections, remove old JS blog buttons
Remove JS-based function-return based blog buttons. While these perform
better on the desktop site, they are a pain to maintain, and the
iframe-based windows are downright simple by comparison. Since the
desktop site is essentially a tech demo anyhow, making the iframes
perfom well everywhere is not my top priority.

Add the keyboards and games blog sections. Both have introduction posts.

Add a review of Ratchet and Clank for the PS2. This is my first
full-form game reivew in the blog, hopefully of many.