Transition to Pelican
Transition site to Pelican. This is the result of the output directory
on my first attempt, still with the default theme and everything.
Eventually I'll update it further to make the desktop site part of
Pelican and therefore include newer blog posts, but it is what it is for
now.